add gard and new routing

+++ repositories/contacts.js
@@ -1,28 +1,28 @@
 const Contact = require('../model/contact')
 
-const listContacts = async () => {
-  const result = await Contact.find()
+const listContacts = async (userId) => {
+  const result = await Contact.find({ owner: userId})
   return result
 }
 
-const getContactById = async (contactId) => {
-  const result = await Contact.findOne({ _id: contactId })
+const getContactById = async (userId, contactId) => {
+  const result = await Contact.findOne({ _id: contactId, owner: userId })
   return result
 }
 
-const removeContact = async (contactId) => {
-  const result = await Contact.findOneAndDelete({_id: contactId})
+const removeContact = async (userId, contactId) => {
+  const result = await Contact.findOneAndDelete({_id: contactId, owner: userId})
   return result
 }
 
-const addContact = async (body) => {
-  const result = await Contact.create(body)
+const addContact = async (userId, body) => {
+  const result = await Contact.create({ owner: userId, ...body })
   return result
 }
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (userId,contactId, body) => {
   const result = await Contact.findOneAndUpdate(
-    {_id: contactId},
+    {_id: contactId, owner: userId},
     { ...body},
     {new: true})
